Move initialization of sensors and actuators to app.py
